Move signing-utils into domain
# Conflicts:
#	packages/wallet/src/redux/channel-store/reducer.ts
