--- packages/wallet/src/domain/get-channel-id.ts
@@ -1,3 +1,0 @@
-import { soliditySha256 } from 'ethers/utils';
-
-export function getChannelId() {}
