+++ packages/wallet/src/domain/index.ts
@@ -1,1 +1,2 @@
 export * from './commitments';
+export * from './signing-utils';
