+++ packages/wallet/src/domain/signing-utils.ts
