+++ packages/wallet/src/redux/channel-store/closing/reducer.ts
@@ -7,7 +7,7 @@ import {
   signCommitment,
   signVerificationData,
   validCommitmentSignature,
-} from '../../../utils/signing-utils';
+} from '../../../domain';
 import {
   closeSuccess,
   concludeSuccess,
