+++ packages/wallet/src/redux/channel-store/running/__tests__/running.test.ts
@@ -9,7 +9,7 @@ import {
   itTransitionsToChannelStateType,
   itSendsThisMessage,
 } from '../../../__tests__/helpers';
-import * as SigningUtil from '../../../../utils/signing-utils';
+import * as SigningUtil from '../../../../domain';
 
 const {
   asAddress,
