+++ packages/wallet/src/redux/channel-store/withdrawing/__tests__/withdrawing.test.ts
@@ -6,7 +6,7 @@ import * as actions from '../../../actions';
 import { itTransitionsToChannelStateType } from '../../../__tests__/helpers';
 import * as scenarios from '../../../__tests__/test-scenarios';
 import * as TransactionGenerator from '../../../../utils/transaction-generator';
-import * as SigningUtil from '../../../../utils/signing-utils';
+import * as SigningUtil from '../../../../domain';
 import { WalletProtocol } from '../../../types';
 
 const {
