+++ packages/wallet/src/redux/protocols/responding/reducer.ts
@@ -13,7 +13,7 @@ import {
   transactionReducer,
 } from '../transaction-submission/reducer';
 import { SharedData } from '../../state';
-import * as SigningUtils from '../../../utils/signing-utils';
+import * as SigningUtils from '../../../domain';
 import { isTransactionAction } from '../transaction-submission/actions';
 import {
   isTerminal,
